implement simple energy dampening
deals with energy drift, such as with larger timesteps where the energy of the system grows over time and the orbits of the bodies grows larger and larger. Fairly rudimentary now, and results in swings between energy being higher than desired and lower than desired.

Possible improvements: calculate energy more frequently to better respond to changes. Energy could be calculated more efficiently as part of the sim updating acceleration for all the bodies, this needs to be done on the n-body lib side. Could monitor direction energy is changing, to avoid applying more damping than needed.